#20 Move select-related tests into separate file and add xfailing test for named select block
