+++ src/fparser/tests/test_parser.py
@@ -95,19 +95,6 @@ def parse(cls, line, label='', isfree=True, isstrict=False):
     raise ValueError('parsing %r with %s pattern failed'%(line, cls.__name__))
 
 
-# We need to monkeypatch the logger used by fparser because it grabs
-# stdout before the pytest framework can intercept it. In python < 3
-# you can't make a lambda out of 'print' because print is not a
-# function (you cannot do "x = print y" for instance). Therefore we
-# have to create our own function that simply wraps print and returns
-# a value.
-def print_wrapper(arg):
-    ''' A wrapper that allows us to call print as a function. Used for
-    monkeypatching logging calls. '''
-    print arg
-    return None
-
-
 def test_assignment():
     assert_equal(parse(Assignment, 'a=b'), 'a = b')
     assert_equal(parse(PointerAssignment,'a=>b'), 'a => b')
@@ -565,231 +552,3 @@ def test_implicit():
                  'IMPLICIT INTEGER ( i-m, p, q-r )')
     assert_equal(parse(Implicit,'implicit integer (i-m), real (z)'),
                  'IMPLICIT INTEGER ( i-m ), REAL ( z )')
-
-
-def test_select_case():
-    '''Test that fparser correctly recognises select case'''
-    from fparser import api
-    import fparser
-    source_str = '''
-    subroutine foo
-    integer :: iflag = 1
-    real    :: aval = 0.0
-    select case(iflag)
-    case(1)
-      aval = 1.0
-    case default
-      aval = 0.0
-    end select
-    ! Same again but with more white space
-    select   case(iflag)
-    case ( 1 )
-      aval = 1.0
-    case  default
-      aval = 0.0
-    end select
-    end subroutine foo
-    '''
-    tree = api.parse(source_str, isfree=True, isstrict=False)
-    assert tree
-    select_list = []
-    for statement in tree.content[0].content:
-        if isinstance(statement, fparser.block_statements.SelectCase):
-            select_list.append(statement)
-    assert len(select_list) == 2
-    for statement in select_list:
-        assert isinstance(statement, fparser.block_statements.SelectCase)
-        assert isinstance(statement.content[0], fparser.statements.Case)
-        assert isinstance(statement.content[2], fparser.statements.Case)
-        assert isinstance(statement.content[3], fparser.statements.Assignment)
-    gen = str(tree)
-    print gen
-    assert "SELECT CASE ( iflag )" in gen
-
-
-def test_select_type():
-    '''Test that fparser correctly recognises select type'''
-    from fparser import api
-    import fparser
-    source_str = '''
-    subroutine foo(an_object)
-    class(*) :: an_object
-    real    :: aval = 0.0
-    select type(an_object)
-    type is (some_type)
-      aval = 1.0
-    class is (some_class)
-      aval = 0.0
-    class default
-      aval = -1.0
-    end select
-    ! Same again but with more white space
-    select   type ( an_object )
-    type  is  ( some_type )
-      aval = 1.0
-    class  is  ( some_class)
-      aval = 0.0
-    class   default
-      aval = -1.0
-    end select
-    end subroutine foo
-    '''
-    tree = api.parse(source_str, isfree=True, isstrict=False)
-    assert tree
-    select_list = []
-    for statement in tree.content[0].content:
-        if isinstance(statement, fparser.block_statements.SelectType):
-            select_list.append(statement)
-    assert len(select_list) == 2
-    for statement in select_list:
-        assert isinstance(statement, fparser.block_statements.SelectType)
-        assert isinstance(statement.content[0], fparser.statements.TypeIs)
-        assert isinstance(statement.content[2], fparser.statements.ClassIs)
-        assert isinstance(statement.content[3], fparser.statements.Assignment)
-        assert isinstance(statement.content[4], fparser.statements.ClassIs)
-    gen = str(tree)
-    print gen
-    assert "SELECT TYPE ( an_object )" in gen
-    assert "TYPE IS ( some_type )" in gen
-    assert "CLASS IS ( some_class )" in gen
-    assert "CLASS DEFAULT" in gen
-
-
-def test_type_is_process_item(monkeypatch, capsys):
-    ''' Test error condition raised in TypeIs.process_item() method '''
-    from fparser import api
-    import fparser
-    source_str = '''
-    subroutine foo(an_object)
-    class(*) :: an_object
-    real    :: aval = 0.0
-    select type(an_object)
-    type is (some_type)
-      aval = 1.0
-    class is (some_class)
-      aval = 0.0
-    end select
-    end subroutine foo
-    '''
-    tree = api.parse(source_str, isfree=True, isstrict=False)
-    assert tree
-    statement = None # Keeps pylint happy
-    for statement in tree.content[0].content:
-        if isinstance(statement, fparser.block_statements.SelectType):
-            break
-    assert isinstance(statement, fparser.block_statements.SelectType)
-    assert isinstance(statement.content[0], fparser.statements.TypeIs)
-    typeis = statement.content[0]
-    typeis.parent.name = "not_a_name"
-    monkeypatch.setattr(typeis.item, "get_line",
-                        lambda: "type is (blah): wrong_name")
-    # Monkeypatch the typeis object so that a call to self.warning
-    # (which normally results in a call to the logger) gets replaced
-    # with a call to our print_wrapper() function
-    monkeypatch.setattr(typeis, "warning", print_wrapper)
-    typeis.process_item()
-    output, _ = capsys.readouterr()
-    print output
-    assert "expected type-is-construct-name 'not_a_name' but got " in output
-
-
-def test_type_is_to_fortran():
-    ''' Test error condition raised in TypeIs.to_fortran() method '''
-    from fparser import api
-    import fparser
-    from fparser.utils import ParseError
-    source_str = '''
-    subroutine foo(an_object)
-    class(*) :: an_object
-    real    :: aval = 0.0
-    select type(an_object)
-    type is (some_type)
-      aval = 1.0
-    class is (some_class)
-      aval = 0.0
-    end select
-    end subroutine foo
-    '''
-    tree = api.parse(source_str, isfree=True, isstrict=False)
-    assert tree
-    statement = None # Keeps pylint happy
-    for statement in tree.content[0].content:
-        if isinstance(statement, fparser.block_statements.SelectType):
-            break
-    assert isinstance(statement, fparser.block_statements.SelectType)
-    assert isinstance(statement.content[0], fparser.statements.TypeIs)
-    typeis = statement.content[0]
-    typeis.name = "some_name"
-    fort = typeis.tofortran()
-    assert "TYPE IS ( some_type ) some_name" in fort
-    # Now break the internal state and check for expected exception
-    typeis.items = None
-    with pytest.raises(ParseError) as excinfo:
-        _ = typeis.tofortran()
-    assert "TYPE IS construct must have arguments" in str(excinfo)
-
-
-def test_class_is_process_item(monkeypatch, capsys):
-    ''' Test error condition raised in ClassIs.process_item() method '''
-    from fparser import api
-    import fparser
-    source_str = '''
-    subroutine foo(an_object)
-    class(*) :: an_object
-    real    :: aval = 0.0
-    select type(an_object)
-    class is (some_class)
-      aval = 0.0
-    class default
-      aval = -1.0
-    end select
-    end subroutine foo
-    '''
-    tree = api.parse(source_str, isfree=True, isstrict=False)
-    assert tree
-    statement = None # Keeps pylint happy
-    for statement in tree.content[0].content:
-        if isinstance(statement, fparser.block_statements.SelectType):
-            break
-    assert isinstance(statement, fparser.block_statements.SelectType)
-    assert isinstance(statement.content[0], fparser.statements.ClassIs)
-    clsis = statement.content[0]
-    clsis.parent.name = "not_a_name"
-    monkeypatch.setattr(clsis.item, "get_line",
-                        lambda: "class is (blah): wrong_name")
-    # Monkeypatch the typeis object so that a call to self.warning
-    # (which normally results in a call to the logger) gets replaced
-    # with a call to our print_wrapper() function
-    monkeypatch.setattr(clsis, "warning", print_wrapper)
-    clsis.process_item()
-    output, _ = capsys.readouterr()
-    print output
-    assert "expected class-construct-name 'not_a_name' but got " in output
-
-
-def test_class_is_to_fortran():
-    ''' Test ClassIs.to_fortran() method '''
-    from fparser import api
-    import fparser
-    source_str = '''
-    subroutine foo(an_object)
-    class(*) :: an_object
-    real    :: aval = 0.0
-    select type(an_object)
-    class is (some_class)
-      aval = 0.0
-    end select
-    end subroutine foo
-    '''
-    tree = api.parse(source_str, isfree=True, isstrict=False)
-    assert tree
-    statement = None # Keeps pylint happy
-    for statement in tree.content[0].content:
-        if isinstance(statement, fparser.block_statements.SelectType):
-            break
-    assert isinstance(statement, fparser.block_statements.SelectType)
-    assert isinstance(statement.content[0], fparser.statements.ClassIs)
-    clsis = statement.content[0]
-    clsis.name = "some_name"
-    fort = clsis.tofortran()
-    assert "CLASS IS ( some_class ) some_name" in fort
